+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 0.9.0
 
+#### 💥 Breaking
+
+- Removed `install_global`, use `cargo install` instead.
+- Removed `uninstall_global`, use `cargo uninstall` instead.
+
 #### 🚀 Updates
 
 - Updated to support proto v0.31 release.